requirements.txt
Pillow==12.1.1
pytest>=8.0.0
